+++ app/src/test/java/com/speakeasy/watsonbarassistant/MainMenuTest.kt
@@ -1,0 +1,18 @@
+package com.speakeasy.watsonbarassistant
+
+import junit.framework.TestCase
+import org.junit.Assert
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class MapMenuTest: TestCase() {
+
+    @Test
+    fun test() {
+        Assert.assertEquals(1, 1)
+    }
+}
